Read instalação e read conta de energia

+++ src/main/java/application/Main.java
@@ -206,9 +206,9 @@ public class Main extends Application {
             n.setEne_consumo_conta_mes(BigInteger.valueOf(Long.parseLong(txtConsumo.getText())));
             n.setEne_valor_total(BigDecimal.valueOf(Double.parseDouble(txtTotalPagar.getText())));
             n.setEne_const_multi(BigDecimal.valueOf(Double.parseDouble(txtConstMulti.getText())));
-            n.setEne_numero_medidor(Integer.parseInt(txtNRmedidor.getText()));
-            n.setEne_leitura_anterior_cod(Integer.parseInt(txtLeituraAnterior.getText()));
-            n.setEne_leitura_atual_cod(Integer.parseInt(txtLeituraAtual.getText()));
+            n.setEne_numero_medidor(BigInteger.valueOf(Long.parseLong(txtNRmedidor.getText())));
+            n.setEne_leitura_anterior_cod(BigInteger.valueOf(Long.parseLong(txtLeituraAnterior.getText())));
+            n.setEne_leitura_atual_cod(BigInteger.valueOf(Long.parseLong(txtLeituraAtual.getText())));
             n.setEne_data_leitura_anterior(txtDataLeituraAnterior.getText());
             n.setEne_data_leitura_atual(txtDataLeituraAtual.getText());
             n.setEne_tipo_bandeira(String.valueOf(comboBandeirasTarifarias.getValue()));
